refactor(directions): rename directions hook file to directions-service
