+++ library/src/index.ts
@@ -1,6 +1,6 @@
 // codegen:start {preset: barrel, include: ./**/*, exclude: [./index.ts, ./types/*, ./google-map.ts]}
 export * from './hooks/autocomplete';
-export * from './hooks/directions';
+export * from './hooks/directions-service';
 export * from './hooks/distance-matrix';
 export * from './hooks/elevation';
 export * from './hooks/geocoder';
